+++ ddb/feature/certs/__init__.py
@@ -18,6 +18,10 @@ class CertsFeature(Feature):
         return "certs"
 
     @property
+    def dependencies(self) -> Iterable[str]:
+        return ["core"]
+
+    @property
     def schema(self) -> ClassVar[FeatureSchema]:
         return CertsSchema