+++ ddb/feature/copy/__init__.py
@@ -18,6 +18,10 @@ class CopyFeature(Feature):
         return "copy"
 
     @property
+    def dependencies(self) -> Iterable[str]:
+        return ["core"]
+
+    @property
     def schema(self) -> ClassVar[FeatureSchema]:
         return CopySchema