+++ ddb/feature/file/__init__.py
@@ -18,6 +18,10 @@ class FileFeature(Feature):
         return "file"
 
     @property
+    def dependencies(self) -> Iterable[str]:
+        return ["core"]
+
+    @property
     def schema(self) -> ClassVar[FeatureSchema]:
         return FileSchema