+++ ddb/feature/gitignore/__init__.py
@@ -17,6 +17,10 @@ class GitignoreFeature(Feature):
         return "gitignore"
 
     @property
+    def dependencies(self) -> Iterable[str]:
+        return ["core"]
+
+    @property
     def schema(self) -> ClassVar[GitignoreSchema]:
         return GitignoreSchema