+++ ddb/feature/symlinks/__init__.py
@@ -22,7 +22,7 @@ class SymlinksFeature(Feature):
 
     @property
     def dependencies(self) -> Iterable[str]:
-        return ["file"]
+        return ["core", "file"]
 
     @property
     def schema(self) -> ClassVar[SymlinksSchema]: