+++ ddb/feature/ytt/__init__.py
@@ -21,7 +21,7 @@ class YttFeature(Feature):
 
     @property
     def dependencies(self) -> Iterable[str]:
-        return ["file"]
+        return ["core", "file"]
 
     @property
     def schema(self) -> ClassVar[YttSchema]: