+++ CHANGELOG.md
@@ -9,7 +9,7 @@ History
 reported in `refs/head`, it will use `refs/remote/origin` to get branche name.
 - Config: Lists merge strategy is now defined as `override` by default.
 - Config: Add support for `merge`/`value` inside configuration to control merge behavior between configuration files.
-
+- Config: Add support for `ddb.<env>.yml` configuration file based on `core.env.current` value.
 
 1.0.9 (2020-09-12)
 ------------------